requirements.txt
streamlit
pandas
langchain
chromadb
transformers
sentence-transformers
llama-cpp-python
faiss-cpu